refactor(web): reorganize into services/ and features/auth/
