+++ web/src/features/auth/ProtectedRoute.tsx
@@ -1,6 +1,6 @@
 import type { ReactNode } from 'react'
 import { Navigate } from 'react-router-dom'
-import { useAuth } from '@/lib/AuthContext'
+import { useAuth } from './AuthContext'
 
 export function ProtectedRoute({ children }: { children: ReactNode }) {
   const { status } = useAuth()
