+++ web/src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { BrowserRouter } from 'react-router-dom'
 import App from './App.tsx'
-import { AuthProvider } from '@/lib/AuthContext'
+import { AuthProvider } from '@/features/auth/AuthContext'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
