+++ web/src/routes/LoginPage.tsx
@@ -1,80 +1,13 @@
-import { useState, type FormEvent } from 'react'
 import { Navigate } from 'react-router-dom'
-import { useAuth } from '@/lib/AuthContext'
+import { useAuth } from '@/features/auth/AuthContext'
+import { LoginForm } from '@/features/auth/LoginForm'
 
 export function LoginPage() {
-  const { status, requestLoginCode, confirmLoginCode } = useAuth()
-  const [step, setStep] = useState<'email' | 'code'>('email')
-  const [email, setEmail] = useState('')
-  const [code, setCode] = useState('')
-  const [error, setError] = useState<string | null>(null)
-  const [submitting, setSubmitting] = useState(false)
+  const { status } = useAuth()
 
   if (status === 'authenticated') {
     return <Navigate to="/" replace />
   }
 
-  async function handleRequestCode(event: FormEvent) {
-    event.preventDefault()
-    setSubmitting(true)
-    setError(null)
-    const result = await requestLoginCode(email)
-    setSubmitting(false)
-    if (result.ok) {
-      setStep('code')
-    } else {
-      setError(result.error)
-    }
-  }
-
-  async function handleConfirmCode(event: FormEvent) {
-    event.preventDefault()
-    setSubmitting(true)
-    setError(null)
-    const result = await confirmLoginCode(code)
-    setSubmitting(false)
-    if (!result.ok) {
-      setError(result.error)
-    }
-  }
-
-  return (
-    <main>
-      <h1>Log in</h1>
-      {step === 'email' ? (
-        <form onSubmit={handleRequestCode}>
-          <label htmlFor="email">Email</label>
-          <input
-            id="email"
-            type="email"
-            autoComplete="email"
-            required
-            value={email}
-            onChange={(event) => setEmail(event.target.value)}
-          />
-          <button type="submit" disabled={submitting}>
-            Send login code
-          </button>
-        </form>
-      ) : (
-        <form onSubmit={handleConfirmCode}>
-          <p>Enter the code sent to {email}</p>
-          <label htmlFor="code">Code</label>
-          <input
-            id="code"
-            type="text"
-            inputMode="numeric"
-            autoComplete="one-time-code"
-            required
-            value={code}
-            onChange={(event) => setCode(event.target.value)}
-          />
-          <button type="submit" disabled={submitting}>
-            Confirm
-          </button>
-        </form>
-      )}
-      {error && <p role="alert">{error}</p>}
-    </main>
-  )
+  return <LoginForm />
 }
